refactor: Refactor layout: move header actions to sidebar
- Removed the Header component and its usage from the dashboard layout.
- Moved the sign out button and theme toggle into the Sidebar for a more consolidated navigation experience.
- Updated the destructive button style for consistency.
- Deleted the Header component and its Storybook story as they are no longer needed.
