+++ src/app/(dashboard)/layout.tsx
@@ -1,5 +1,4 @@
 import Sidebar from "@/components/molecules/Sidebar";
-import Header from "@/components/molecules/Header";
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
   return (
@@ -7,7 +6,6 @@ export default function DashboardLayout({ children }: { children: React.ReactNod
       <Sidebar />
 
       <div className="flex flex-1 flex-col">
-        <Header title="Dashboard" />
 
         <main className="flex-1 p-6">{children}</main>
       </div>
