--- src/components/molecules/Header.tsx
@@ -1,30 +1,0 @@
-"use client";
-
-import { signOut } from "next-auth/react";
-import ThemeToggle from "./ThemeToggle";
-
-export default function Header({ title = "Dashboard" }: { title?: string }) {
-  const handleSignOut = async () => {
-    await signOut({
-      callbackUrl: "/login",
-      redirect: true,
-    });
-  };
-
-  return (
-    <header className="border-b border-zinc-200 bg-white px-6 py-4 dark:border-zinc-800 dark:bg-zinc-900">
-      <div className="flex items-center justify-between">
-        <h2 className="text-lg font-semibold text-zinc-900 dark:text-zinc-50">{title}</h2>
-        <div className="flex items-center gap-4">
-          <ThemeToggle />
-          <button
-            onClick={handleSignOut}
-            className="text-sm text-zinc-600 hover:text-zinc-900 dark:text-zinc-400 dark:hover:text-zinc-50 transition-colors"
-          >
-            Sign out
-          </button>
-        </div>
-      </div>
-    </header>
-  );
-}
